+++ openMP_backprop.c
@@ -8,10 +8,10 @@
 #include "openMP_backprop.h"
 
 #define CORE_NUM 56
-void printOutError(float* error, int lenght)
+void printOutError(double* error, int lenght)
 {
   assert(error);
-  float sum = 0.0;
+  double sum = 0.0;
   for (int i = 0; i < lenght; ++i)
     {
       sum += error[i];
@@ -19,7 +19,7 @@ void printOutError(float* error, int lenght)
   printf("The total error is %f\n", sum);
 }
 
-void calerrorOuputO(float* output, float* truth, float* res, int length)
+void calerrorOuputO(double* output, double* truth, double* res, int length)
 {
   assert(output);
   assert(truth);
@@ -38,12 +38,12 @@ void calerrorOuputO(float* output, float* truth, float* res, int length)
 // derivative of error with respect to input of next layer as well as length of
 // the current node and length of the next node
 void calerrorOuput(
-    float* weightCur, float* nexterrorVAl, float* res, int lenCur, int lenNex)
+    double* weightCur, double* nexterrorVAl, double* res, int lenCur, int lenNex)
 {
   assert(nexterrorVAl);
   assert(res);
   assert(weightCur);
-  float* temp = NULL;
+  double* temp = NULL;
   temp        = transposeMatrix(weightCur, lenNex, lenCur);
   matrixMultiplication(temp, nexterrorVAl, res, lenCur, lenNex, 1);
   free(temp);
@@ -51,7 +51,7 @@ void calerrorOuput(
 
 // calculate the derivative of error w.r.t to value of a node
 // parameters: the current layer value, derivative of error w.r.t output
-void calerrorVal(float* layerVal, float* errorOutput, float* res, int length)
+void calerrorVal(double* layerVal, double* errorOutput, double* res, int length)
 {
   assert(layerVal);
   assert(errorOutput);
@@ -70,7 +70,7 @@ void calerrorVal(float* layerVal, float* errorOutput, float* res, int length)
 // calculate the gradient descent
 // parameters: output of previous layer, derivative of error w.r.t value
 void calgrad(
-    float* curOutput, float* errorVal, float* res, int lenNext, int lenCur)
+    double* curOutput, double* errorVal, double* res, int lenNext, int lenCur)
 {
   assert(curOutput);
   assert(errorVal);
@@ -97,9 +97,9 @@ void backpropAuto(AUTOW*   autoweights,
                   int      stage)
 {
   // temporary variables for storing numbers between layers
-  float* derErrorOutput0 = NULL;
-  float* derErrorOutput  = NULL;
-  float* derErrorVal     = NULL;
+  double* derErrorOutput0 = NULL;
+  double* derErrorOutput  = NULL;
+  double* derErrorVal     = NULL;
   switch (stage)
     {
       case 1:
@@ -115,9 +115,9 @@ void backpropAuto(AUTOW*   autoweights,
           }
 #endif
         // train decoder layer output
-        derErrorOutput0 = (float*)malloc(773 * sizeof(float));
-        calerrorOuputO((float*)(decodelayer->output),
-                       (float*)(autolayer->input),
+        derErrorOutput0 = (double*)malloc(773 * sizeof(double));
+        calerrorOuputO((double*)(decodelayer->output),
+                       (double*)(autolayer->input),
                        derErrorOutput0,
                        773);
 #ifdef DEBUG
@@ -128,17 +128,17 @@ void backpropAuto(AUTOW*   autoweights,
         // printOutError(derErrorOutput0, 773);
 
 #endif
-        derErrorVal = (float*)malloc(773 * sizeof(float));
+        derErrorVal = (double*)malloc(773 * sizeof(double));
         calerrorVal(
-            (float*)decodelayer->output, derErrorOutput0, derErrorVal, 773);
+            (double*)decodelayer->output, derErrorOutput0, derErrorVal, 773);
         for (int l = 0; l < 10; ++l)
           {
             printf("\nerror val: %f ", derErrorVal[l]);
           }
         printOutError(derErrorOutput0, 773);
-        calgrad((float*)decodelayer->layer3,
+        calgrad((double*)decodelayer->layer3,
                 derErrorVal,
-                (float*)decodegrad->weight3,
+                (double*)decodegrad->weight3,
                 773,
                 600);
 
@@ -148,8 +148,8 @@ void backpropAuto(AUTOW*   autoweights,
           }
 
         // train input layer of encoder
-        derErrorOutput = (float*)malloc(600 * sizeof(float));
-        calerrorOuput((float*)autoweights->weight0,
+        derErrorOutput = (double*)malloc(600 * sizeof(double));
+        calerrorOuput((double*)autoweights->weight0,
                       derErrorVal,
                       derErrorOutput,
                       600,
@@ -160,15 +160,15 @@ void backpropAuto(AUTOW*   autoweights,
           }
         free(derErrorVal);
         derErrorVal = NULL;
-        derErrorVal = (float*)malloc(600 * sizeof(float));
-        calerrorVal((float*)autolayer->input, derErrorOutput, derErrorVal, 600);
+        derErrorVal = (double*)malloc(600 * sizeof(double));
+        calerrorVal((double*)autolayer->input, derErrorOutput, derErrorVal, 600);
         for (int l = 0; l < 10; ++l)
           {
             printf("\n error val %f ", derErrorVal[l]);
           }
-        calgrad((float*)autolayer->layer1,
+        calgrad((double*)autolayer->layer1,
                 derErrorVal,
-                (float*)autograd->weight0,
+                (double*)autograd->weight0,
                 600,
                 773);
         for (int l = 0; l < 10; ++l)
@@ -183,60 +183,60 @@ void backpropAuto(AUTOW*   autoweights,
         break;
       case 2:
         // train decoder layer output
-        derErrorOutput0 = (float*)malloc(773 * sizeof(float));
-        calerrorOuputO((float*)decodelayer->output,
-                       (float*)autolayer->input,
+        derErrorOutput0 = (double*)malloc(773 * sizeof(double));
+        calerrorOuputO((double*)decodelayer->output,
+                       (double*)autolayer->input,
                        derErrorOutput0,
                        773);
-        derErrorVal = (float*)malloc(773 * sizeof(float));
+        derErrorVal = (double*)malloc(773 * sizeof(double));
         calerrorVal(
-            (float*)decodelayer->output, derErrorOutput0, derErrorVal, 773);
+            (double*)decodelayer->output, derErrorOutput0, derErrorVal, 773);
 
         // train layer 3 of decoder
-        derErrorOutput = (float*)malloc(400 * sizeof(float));
-        calerrorOuput((float*)autoweights->weight1,
+        derErrorOutput = (double*)malloc(400 * sizeof(double));
+        calerrorOuput((double*)autoweights->weight1,
                       derErrorVal,
                       derErrorOutput,
                       400,
                       600);
         free(derErrorVal);
         derErrorVal = NULL;
-        derErrorVal = (float*)malloc(600 * sizeof(float));
+        derErrorVal = (double*)malloc(600 * sizeof(double));
         calerrorVal(
-            (float*)autolayer->layer1, derErrorOutput, derErrorVal, 400);
+            (double*)autolayer->layer1, derErrorOutput, derErrorVal, 400);
         calgrad(
-            derErrorOutput0, derErrorVal, (float*)autograd->weight1, 400, 600);
+            derErrorOutput0, derErrorVal, (double*)autograd->weight1, 400, 600);
         free(derErrorOutput);
 
         // train layer 1 of encoder
-        derErrorOutput = (float*)malloc(400 * sizeof(float));
-        calerrorOuput((float*)autoweights->weight1,
+        derErrorOutput = (double*)malloc(400 * sizeof(double));
+        calerrorOuput((double*)autoweights->weight1,
                       derErrorVal,
                       derErrorOutput,
                       400,
                       600);
         free(derErrorVal);
         derErrorVal = NULL;
-        derErrorVal = (float*)malloc(600 * sizeof(float));
+        derErrorVal = (double*)malloc(600 * sizeof(double));
         calerrorVal(
-            (float*)autolayer->layer1, derErrorOutput, derErrorVal, 400);
+            (double*)autolayer->layer1, derErrorOutput, derErrorVal, 400);
         calgrad(
-            derErrorOutput0, derErrorVal, (float*)autograd->weight1, 400, 600);
+            derErrorOutput0, derErrorVal, (double*)autograd->weight1, 400, 600);
 
         // train input layer of encoder
         free(derErrorOutput);
-        derErrorOutput = (float*)malloc(600 * sizeof(float));
-        calerrorOuput((float*)autoweights->weight0,
+        derErrorOutput = (double*)malloc(600 * sizeof(double));
+        calerrorOuput((double*)autoweights->weight0,
                       derErrorVal,
                       derErrorOutput,
                       600,
                       773);
         free(derErrorVal);
         derErrorVal = NULL;
-        derErrorVal = (float*)malloc(600 * sizeof(float));
-        calerrorVal((float*)autolayer->input, derErrorOutput, derErrorVal, 600);
+        derErrorVal = (double*)malloc(600 * sizeof(double));
+        calerrorVal((double*)autolayer->input, derErrorOutput, derErrorVal, 600);
         calgrad(
-            derErrorOutput0, derErrorVal, (float*)autograd->weight0, 600, 773);
+            derErrorOutput0, derErrorVal, (double*)autograd->weight0, 600, 773);
 
         free(derErrorOutput0);
         free(derErrorOutput);
@@ -245,20 +245,20 @@ void backpropAuto(AUTOW*   autoweights,
         break;
       // case 3:
       //   // train layer 2 of decoder
-      //   derErrorOutput0 = (float*)malloc(400 * sizeof(float));
+      //   derErrorOutput0 = (double*)malloc(400 * sizeof(double));
       //   calerrorOuputO(
       //       decodelayer->layer2, autolayer->layer2, derErrorOutput0, 400);
-      //   derErrorVal = (float*)malloc(400 * sizeof(float));
+      //   derErrorVal = (double*)malloc(400 * sizeof(double));
       //   calerrorVal(decodelayer->layer2, derErrorOutput0, derErrorVal, 400);
       //   calgrad(decodelayer->layer1, derErrorVal, decodegrad->layer2, 400,
       //   200);
 
-      //   derErrorOutput = (float*)malloc(200 * sizeof(float));
+      //   derErrorOutput = (double*)malloc(200 * sizeof(double));
       //   calerrorOuput(
       //       autoweights->weight3, derErrorVal, derErrorOutput, 200, 400);
       //   free(derErrorVal);
       //   derErrorVal = NULL;
-      //   derErrorVal = (float*)malloc(200 * sizeof(float));
+      //   derErrorVal = (double*)malloc(200 * sizeof(double));
       //   calerrorVal(autolayer->layer2, derErrorOutput, derErrorVal, 200);
       //   calgrad(derErrorOutput0, derErrorVal, autograd->weight2, 200, 400);
 
@@ -268,20 +268,20 @@ void backpropAuto(AUTOW*   autoweights,
       //   break;
       // case 4:
       //   // train layer 1 of decoder
-      //   derErrorOutput0 = (float*)malloc(200 * sizeof(float));
+      //   derErrorOutput0 = (double*)malloc(200 * sizeof(double));
       //   calerrorOuputO(
       //       decodelayer->layer1, autolayer->layer3, derErrorOutput0, 200);
-      //   derErrorVal = (float*)malloc(200 * sizeof(float));
+      //   derErrorVal = (double*)malloc(200 * sizeof(double));
       //   calerrorVal(decodelayer->layer1, derErrorOutput0, derErrorVal, 200);
       //   calgrad(autolayer->output, derErrorVal, decodegrad->layer1, 200,
       //   100);
 
-      //   derErrorOutput = (float*)malloc(100 * sizeof(float));
+      //   derErrorOutput = (double*)malloc(100 * sizeof(double));
       //   calerrorOuput(
       //       autoweights->weight3, derErrorVal, derErrorOutput, 100, 200);
       //   free(derErrorVal);
       //   derErrorVal = NULL;
-      //   derErrorVal = (float*)malloc(100 * sizeof(float));
+      //   derErrorVal = (double*)malloc(100 * sizeof(double));
       //   calerrorVal(autolayer->layer3, derErrorOutput, derErrorVal, 100);
       //   calgrad(derErrorOutput0, derErrorVal, autograd->weight3, 100, 200);
